requirements.txt
albumentations==1.2.1
ipython==7.34.0
ipywidgets==7.7.1
matplotlib==3.7.1
numpy==1.22.4
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
plotly==5.13.1
protobuf==3.20.3
scikit_learn==1.2.2
seaborn==0.12.2
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
wandb==0.15.0